Monolith Projects has been divided in two parts. This Microservice was built for getting data from Central Bank of Azerbaijan Republic,in addition, it can convert data which was getting .xml format to json format.

+++ src/main/java/com/example/mscbarconfig/model/CurrencyRate.java
@@ -2,6 +2,8 @@ package com.example.mscbarconfig.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 import java.util.Objects;
 
 public class CurrencyRate {
@@ -12,12 +14,12 @@ public class CurrencyRate {
     @JsonProperty("Name")
     private String name;
     @JsonProperty("Value")
-    private Double value;
+    private BigDecimal value;
 
     public CurrencyRate() {
     }
 
-    public CurrencyRate(String code, String nominal, String name, Double value) {
+    public CurrencyRate(String code, String nominal, String name, BigDecimal value) {
         this.code = code;
         this.nominal = nominal;
         this.name = name;
@@ -48,11 +50,11 @@ public class CurrencyRate {
         this.name = name;
     }
 
-    public Double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 
-    public void setValue(Double value) {
+    public void setValue(BigDecimal value) {
         this.value = value;
     }
